requirements.txt
fastapi
requests
beautifulsoup4
sentence-transformers
python-dotenv
uvicorn